+++ flexmeasures/ui/crud/users.py
@@ -75,10 +75,28 @@ def process_internal_api_response(
     return user_data
 
 
+def get_users_by_account(
+    account_id: int | str, include_inactive: bool = False
+) -> list[User]:
+    get_users_response = InternalApi().get(
+        url_for(
+            "UserAPI:index",
+            account_id=account_id,
+            include_inactive=include_inactive,
+        )
+    )
+    users = [
+        process_internal_api_response(user, make_obj=True)
+        for user in get_users_response.json()
+    ]
+    return users
+
+
 class UserCrudUI(FlaskView):
     route_base = "/users"
     trailing_slash = False
 
+    @login_required
     def index(self):
         """/users"""
         include_inactive = request.args.get("include_inactive", "0") != "0"
@@ -90,21 +108,12 @@ class UserCrudUI(FlaskView):
         else:
             accounts = [current_user.account]
         for account in accounts:
-            get_users_response = InternalApi().get(
-                url_for(
-                    "UserAPI:index",
-                    account_id=account.id,
-                    include_inactive=include_inactive,
-                )
-            )
-            users += [
-                process_internal_api_response(user, make_obj=True)
-                for user in get_users_response.json()
-            ]
+            users += get_users_by_account(account.id, include_inactive)
         return render_flexmeasures_template(
             "crud/users.html", users=users, include_inactive=include_inactive
         )
 
+    @login_required
     @roles_accepted(ADMIN_ROLE, ADMIN_READER_ROLE)
     def get(self, id: str):
         """GET from /users/<id>"""